pyproject.toml
[tool.poetry]
name = "poker-fr-tournaments"
version = "0.1.0"
description = "Site de référence des tournois de poker en France"
authors = ["{{USER_EMAIL}}"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.110.0"
uvicorn = "0.28.0"
sqlalchemy = "2.0.23"
psycopg2-binary = "2.9.9"
pydantic = "^2.6.3"
python-dateutil = "^2.8.2"
feedparser = "^6.0.11"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
PyYAML = "^6.0.1"
geopy = "^2.4.1"
ics = "^0.7.2"
httpx = "^0.27.0"
ics-parser = "0.0.1" # dummy placeholder; optional

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.1.3"
black = "^23.9.1"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"